+++ Source/SwiftLintFramework/Configuration.swift
@@ -159,6 +159,8 @@ public struct Configuration {
         }
         rules.append(NestingRule())
         rules.append(ControlStatementRule())
+        rules.append(OpeningBraceRule())
+
         return rules
     }